use BigDecimal and check if transferred amount is positive

+++ src/main/java/fun/wackloner/transferwiser/service/AccountService.java
@@ -6,6 +6,7 @@ import fun.wackloner.transferwiser.repository.AccountRepository;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -41,21 +42,21 @@ public class AccountService {
 
     @POST
     @Path("/{id}/deposit")
-    public Response depositAccount(@PathParam("id") long id, @QueryParam("amount") double amount) {
+    public Response depositAccount(@PathParam("id") long id, @QueryParam("amount") BigDecimal amount) {
         accountRepository.getAccount(id).deposit(amount);
         return Response.noContent().build();
     }
 
     @POST
     @Path("/{id}/withdraw")
-    public Response withdrawAccount(@PathParam("id") long id, @QueryParam("amount") double amount) {
+    public Response withdrawAccount(@PathParam("id") long id, @QueryParam("amount") BigDecimal amount) {
         accountRepository.getAccount(id).withdraw(amount);
         return Response.noContent().build();
     }
 
     @POST
     @Path("/{id}/transfer")
-    public Response transfer(@PathParam("id") long id, @QueryParam("to") long to, @QueryParam("amount") double amount) {
+    public Response transfer(@PathParam("id") long id, @QueryParam("to") long to, @QueryParam("amount") BigDecimal amount) {
         // TODO: check input validity
 
         synchronized (TRANSFER_LOCK) {
